+++ packages/nimble-components/src/select/tests/select.spec.ts
@@ -868,14 +868,19 @@ describe('Select', () => {
             expect(currentSelection?.selected).toBeTrue();
         });
 
-        it('filtering out current selected item and then pressing <Enter> changes value and closes popup', async () => {
+        it('filtering out current selected item and then pressing <Enter> changes value, emits change event, and closes popup', async () => {
             const currentSelection = pageObject.getSelectedOption();
             expect(currentSelection?.text).toBe('One');
             expect(element.value).toBe('one');
 
+            const changeEvent = jasmine.createSpy();
+            element.addEventListener('change', changeEvent);
             await pageObject.openAndSetFilterText('T'); // Matches 'Two' and 'Three'
             pageObject.pressEnterKey();
+            await waitForUpdatesAsync();
             expect(element.value).toBe('two'); // 'Two' is first option in list so it should be selected now
+            expect(pageObject.getDisplayText()).toBe('Two');
+            expect(changeEvent.calls.count()).toBe(1);
             expect(element.open).toBeFalse();
         });
 
@@ -890,14 +895,19 @@ describe('Select', () => {
             expect(element.open).toBeFalse();
         });
 
-        it('filtering out current selected item and then clicking active option changes value and closes popup', async () => {
+        it('filtering out current selected item and then clicking active option changes value, emits change event, and closes popup', async () => {
             const currentSelection = pageObject.getSelectedOption();
             expect(currentSelection?.text).toBe('One');
             expect(element.value).toBe('one');
 
+            const changeEvent = jasmine.createSpy();
+            element.addEventListener('change', changeEvent);
             await pageObject.openAndSetFilterText('T'); // Matches 'Two' and 'Three'
             pageObject.clickActiveItem();
+            await waitForUpdatesAsync();
             expect(element.value).toBe('two'); // 'Two' is first option in list so it should be selected now
+            expect(pageObject.getDisplayText()).toBe('Two');
+            expect(changeEvent.calls.count()).toBe(1);
             expect(element.open).toBeFalse();
         });